Submit problem 2, session 9 homework
Yes. It is much easier. I think I discovered this last time. By renaming
things from "states" to booleans that represent which number is being
used, and using the mostrecent operator, it is easy to see how my
homework from the last time meets the problem 2 specification already.
Just renamed variables and simplified a couple of things to make it more
readable.